+++ src/modules/search/services/index.ts
@@ -1,0 +1,1 @@
+export { KeywordSearchService } from './KeywordSearchService';